Separated classes in Ride into multiple files, standardized ScheduleBar
