+++ lib/models/Ride.dart
@@ -1,13 +1,9 @@
 import 'dart:core';
-import 'package:carriage_rider/pages/Current.dart';
 import 'package:carriage_rider/providers/RiderProvider.dart';
-import 'dart:math';
-import 'package:carriage_rider/pages/RidePage.dart';
 import 'package:flutter/material.dart';
 import 'package:humanize/humanize.dart';
 import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
-import 'package:url_launcher/url_launcher.dart' as UrlLauncher;
 import 'package:carriage_rider/models/Driver.dart';
 import '../utils/CarriageTheme.dart';
 
@@ -321,450 +317,3 @@ T getOrNull<T>(Map<String, dynamic> map, String key, {T parse(dynamic s)}) {
   return parse(x);
 }
 
-class RideCard extends StatelessWidget {
-  RideCard(this.ride,
-      {@required this.showConfirmation,
-        @required this.showCallDriver,
-        @required this.showArrow,
-        this.parentRideID});
-
-  final Ride ride;
-  final bool showConfirmation;
-  final bool showCallDriver;
-  final bool showArrow;
-  final String parentRideID;
-
-  final confirmationStyle = TextStyle(
-    fontWeight: FontWeight.w500,
-    fontSize: 10,
-  );
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: () {
-        Navigator.push(
-            context,
-            new MaterialPageRoute(
-                builder: (context) =>
-                    RidePage(ride, parentRideID: parentRideID)));
-      },
-      child: Container(
-        margin: EdgeInsets.all(2),
-        decoration: CarriageTheme.cardDecoration,
-        child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 24, vertical: 16),
-          child: Row(
-            children: [
-              Expanded(
-                child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: <Widget>[
-                      showConfirmation
-                          ? (ride.type == 'active'
-                          ? Text('Ride Confirmed',
-                          style: confirmationStyle.copyWith(
-                              color: Color(0xFF4CAF50)))
-                          : Text('Ride Requested',
-                          style: confirmationStyle.copyWith(
-                              color: Color(0xFFFF9800))))
-                          : Container(),
-                      SizedBox(height: 4),
-                      ride.buildStartTime(),
-                      SizedBox(height: 16),
-                      Text('From', style: CarriageTheme.directionStyle),
-                      Text(ride.startLocation,
-                          style: CarriageTheme.rideInfoStyle),
-                      SizedBox(height: 8),
-                      Text('To', style: CarriageTheme.directionStyle),
-                      Text(ride.endLocation,
-                          style: CarriageTheme.rideInfoStyle),
-                      SizedBox(height: 16),
-                      showCallDriver
-                          ? Row(
-                        children: <Widget>[
-                          GestureDetector(
-                            onTap: () =>
-                                UrlLauncher.launch('tel://${ride.driver.phoneNumber}'),
-                            child: Container(
-                                decoration: BoxDecoration(
-                                    borderRadius:
-                                    BorderRadius.circular(100),
-                                    border: Border.all(
-                                        width: 0.5,
-                                        color: Colors.black
-                                            .withOpacity(0.25))),
-                                child: Padding(
-                                  padding: const EdgeInsets.all(5),
-                                  child: Icon(Icons.phone,
-                                      size: 20, color: Color(0xFF9B9B9B)),
-                                )),
-                          ),
-                          SizedBox(width: 8),
-                          Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: <Widget>[
-                              Text('Driver',
-                                  style: TextStyle(fontSize: 11)),
-                              Text(
-                                  ride.type == 'active'
-                                      ? 'Confirmed'
-                                      : 'TBD',
-                                  style: CarriageTheme.rideInfoStyle)
-                            ],
-                          )
-                        ],
-                      )
-                          : Container()
-                    ]),
-              ),
-              showArrow ? Icon(Icons.chevron_right, size: 28) : Container()
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class RecurringRidesGenerator {
-  RecurringRidesGenerator(this.originalRides);
-
-  List<Ride> originalRides;
-
-  int daysUntilWeekday(DateTime start, int weekday) {
-    int startWeekday = start.weekday;
-    if (weekday < startWeekday) {
-      weekday += 7;
-    }
-    return weekday - startWeekday;
-  }
-
-  DateTime nextDateOnWeekday(DateTime start, int weekday) {
-    int daysUntil = daysUntilWeekday(start, weekday);
-    return start.add(Duration(days: daysUntil));
-  }
-
-  bool ridesEqualTimeLoc(Ride ride, Ride otherRide) {
-    return (ride.startTime.isAtSameMomentAs(otherRide.startTime) &&
-        ride.endTime.isAtSameMomentAs(otherRide.endTime) &&
-        ride.startLocation == otherRide.startLocation &&
-        ride.endLocation == otherRide.endLocation &&
-        ride.startAddress == otherRide.startAddress &&
-        ride.endAddress == otherRide.endAddress);
-  }
-
-  bool wasDeleted(Ride generatedRide, List<Ride> deletedRides) {
-    return deletedRides
-        .where((deletedRide) => ridesEqualTimeLoc(deletedRide, generatedRide))
-        .isNotEmpty;
-  }
-
-  List<Ride> generateRideInstances() {
-    List<Ride> allRides = [];
-    Map<String, Ride> originalRidesByID = Map();
-    Map<Ride, String> rideInstanceParentIDs = Map();
-
-    for (Ride originalRide in originalRides) {
-      originalRidesByID[originalRide.id] = originalRide;
-    }
-    for (Ride originalRide in originalRides) {
-      // add one time rides
-      if (!originalRide.deleted && !originalRide.recurring) {
-        allRides.add(originalRide);
-      }
-
-      if (originalRide.recurring) {
-        Duration rideDuration =
-        originalRide.endTime.difference(originalRide.startTime);
-        List<Ride> deletedInstances = originalRide.edits
-            .map((rideID) => originalRidesByID[rideID])
-            .where((ride) => ride.deleted)
-            .toList();
-        List<int> days = originalRide.recurringDays;
-
-        // find first occurrence
-        int daysUntilFirstOccurrence = days
-            .map((day) => daysUntilWeekday(originalRide.startTime, day))
-            .reduce(min);
-        DateTime rideStart = originalRide.startTime
-            .add(Duration(days: daysUntilFirstOccurrence));
-        int dayIndex = days.indexOf(rideStart.weekday);
-
-        // generate instances of recurring rides
-        while (rideStart.isBefore(originalRide.endDate) ||
-            rideStart.isAtSameMomentAs(originalRide.endDate)) {
-          // create the new ride and keep track of its parent
-          Ride rideInstance = Ride(
-              id: CarriageTheme.generatedRideID,
-              startLocation: originalRide.startLocation,
-              startAddress: originalRide.startAddress,
-              endLocation: originalRide.endLocation,
-              endAddress: originalRide.endAddress,
-              startTime: rideStart,
-              endTime: rideStart.add(rideDuration));
-          rideInstanceParentIDs[rideInstance] = originalRide.id;
-
-          // if all ride info is equal to a deleted ride, don't add it because it was edited
-          if (!wasDeleted(rideInstance, deletedInstances)) {
-            allRides.add(rideInstance);
-          }
-
-          // find the next occurrence
-          dayIndex = dayIndex == days.length - 1 ? 0 : dayIndex + 1;
-          int daysUntilNextOccurrence =
-          daysUntilWeekday(rideStart, days[dayIndex]);
-          rideStart = rideStart.add(Duration(days: daysUntilNextOccurrence));
-        }
-      }
-    }
-    return allRides;
-  }
-
-  List<Ride> upcomingRides() {
-    List<Ride> allRides = generateRideInstances();
-    return allRides
-        .where((ride) => ride.startTime.isAfter(DateTime.now()))
-        .toList()
-      ..sort((ride1, ride2) =>
-      ride1.startTime.isBefore(ride2.startTime) ? -1 : 1);
-  }
-
-  ListView buildUpcomingRidesList() {
-    List futureRides = upcomingRides();
-    return ListView.separated(
-      shrinkWrap: true,
-      physics: NeverScrollableScrollPhysics(),
-      itemCount: futureRides.length,
-      itemBuilder: (context, index) {
-        return RideCard(
-          futureRides[index],
-          showConfirmation: true,
-          showCallDriver: false,
-          showArrow: true,
-        );
-      },
-      separatorBuilder: (context, index) {
-        return SizedBox(height: 16);
-      },
-    );
-  }
-
-  ListView buildPastRidesList() {
-    List<Ride> allRides = generateRideInstances()
-      ..sort(
-              (ride1, ride2) => ride1.startTime.isBefore(ride2.startTime) ? 1 : -1);
-    return ListView.separated(
-      shrinkWrap: true,
-      physics: NeverScrollableScrollPhysics(),
-      itemCount: allRides.length,
-      itemBuilder: (context, index) {
-        return RideCard(
-          allRides[index],
-          showConfirmation: false,
-          showCallDriver: false,
-          showArrow: true,
-        );
-      },
-      separatorBuilder: (context, index) {
-        return SizedBox(height: 16);
-      },
-    );
-  }
-}
-
-Widget onTheWayRide(context, Ride ride) {
-  String startLocation = ride.startLocation;
-  DateTime startTime = ride.startTime;
-  String timeString = DateFormat.jm().format(startTime);
-  return Row(children: <Widget>[
-    Expanded(
-      child: RichText(
-        text: new TextSpan(
-          style: new TextStyle(
-            fontSize: 20.0,
-            color: Colors.black,
-          ),
-          children: <TextSpan>[
-            new TextSpan(text: 'Head to the '),
-            new TextSpan(
-                text: 'pickup location\n',
-                style: new TextStyle(fontWeight: FontWeight.bold)),
-            new TextSpan(text: 'for @$startLocation by $timeString'),
-          ],
-        ),
-      ),
-    )
-  ]);
-}
-
-Widget arrivedRide(context, Ride ride) {
-  String startLocation = ride.startLocation;
-  return Row(children: <Widget>[
-    Expanded(
-        child: RichText(
-          text: new TextSpan(
-            style: new TextStyle(
-              fontSize: 20.0,
-              color: Colors.black,
-            ),
-            children: <TextSpan>[
-              new TextSpan(text: 'Meet your driver '),
-              new TextSpan(
-                  text: '@$startLocation',
-                  style: new TextStyle(fontWeight: FontWeight.bold)),
-            ],
-          ),
-        ))
-  ]);
-}
-
-Widget pickedUpRide(context, Ride ride) {
-  String endLocation = ride.endLocation;
-  return Row(children: <Widget>[
-    Expanded(
-        child: RichText(
-          text: new TextSpan(
-            style: new TextStyle(
-              fontSize: 20.0,
-              color: Colors.black,
-            ),
-            children: <TextSpan>[
-              new TextSpan(text: 'Your driver will drop you off \n'),
-              new TextSpan(
-                  text: '@$endLocation',
-                  style: new TextStyle(fontWeight: FontWeight.bold)),
-            ],
-          ),
-        ))
-  ]);
-}
-
-Widget completeRide(context) {
-  return Row(children: <Widget>[
-    Expanded(
-        child: RichText(
-          text: new TextSpan(
-            style: new TextStyle(
-              fontSize: 20.0,
-              color: Colors.black,
-            ),
-            children: <TextSpan>[
-              new TextSpan(text: 'Your ride is '),
-              new TextSpan(
-                  text: 'complete!',
-                  style: new TextStyle(fontWeight: FontWeight.bold)),
-            ],
-          ),
-        ))
-  ]);
-}
-
-Widget currentCardInstruction(context, Ride ride) {
-  return ride.status == RideStatus.NOT_STARTED ||
-      ride.status == RideStatus.ON_THE_WAY
-      ? onTheWayRide(context, ride)
-      : ride.status == RideStatus.ARRIVED
-      ? arrivedRide(context, ride)
-      : ride.status == RideStatus.PICKED_UP
-      ? pickedUpRide(context, ride)
-      : completeRide(context);
-}
-
-class CurrentRideCard extends StatelessWidget {
-  CurrentRideCard(
-      this.ride, {
-        @required this.showCallDriver,
-      });
-
-  final Ride ride;
-  final bool showCallDriver;
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: () {
-        if (ride != null) {
-          Navigator.push(context,
-              new MaterialPageRoute(builder: (context) => Current(ride)));
-        }
-      },
-      child: Container(
-        margin: EdgeInsets.all(2),
-        decoration: CarriageTheme.cardDecoration,
-        child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 24, vertical: 16),
-          child: Row(
-            children: [
-              ride == null
-                  ? Expanded(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.center,
-                  children: <Widget>[
-                    SizedBox(height: 20),
-                    Icon(
-                      Icons.directions_car_rounded,
-                      size: 32,
-                      color: Colors.grey,
-                    ),
-                    SizedBox(height: 10),
-                    Text('No current ride',
-                        style: CarriageTheme.body
-                            .copyWith(color: Colors.grey)),
-                    SizedBox(height: 20),
-                  ],
-                ),
-              )
-                  : Expanded(
-                child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: <Widget>[
-                      SizedBox(height: 10),
-                      currentCardInstruction(context, ride),
-                      SizedBox(height: 15),
-                      showCallDriver
-                          ? Row(
-                        children: <Widget>[
-                          GestureDetector(
-                            onTap: () => UrlLauncher.launch(
-                                'tel://${ride.driver.phoneNumber}'),
-                            child: Container(
-                                decoration: BoxDecoration(
-                                    borderRadius:
-                                    BorderRadius.circular(100),
-                                    border: Border.all(
-                                        width: 0.5,
-                                        color: Colors.black
-                                            .withOpacity(0.25))),
-                                child: Padding(
-                                  padding: const EdgeInsets.all(5),
-                                  child: Icon(Icons.phone,
-                                      size: 20,
-                                      color: Color(0xFF4CAF50)),
-                                )),
-                          ),
-                          SizedBox(width: 8),
-                          Column(
-                            crossAxisAlignment:
-                            CrossAxisAlignment.start,
-                            children: <Widget>[
-                              Text('Driver',
-                                  style: TextStyle(fontSize: 11)),
-                              Text(ride.driver.fullName(),
-                                  style:
-                                  CarriageTheme.rideInfoStyle)
-                            ],
-                          )
-                        ],
-                      )
-                          : Container(),
-                      SizedBox(height: 10),
-                    ]),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
