+++ lib/pages/Notifications.dart
@@ -1,5 +1,5 @@
 import 'dart:ui';
-import 'package:carriage_rider/pages/RidePage.dart';
+import 'package:carriage_rider/widgets/ScheduleBar.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:carriage_rider/utils/CarriageTheme.dart';
@@ -15,17 +15,7 @@ class _NotificationsState extends State<Notifications> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        title: PageTitle(title: 'Schedule'),
-        backgroundColor: Colors.white,
-        titleSpacing: 0.0,
-        iconTheme: IconThemeData(color: Colors.black),
-        automaticallyImplyLeading: true,
-        leading: IconButton(
-          icon: Icon(Icons.arrow_back_ios),
-          onPressed: () => Navigator.pop(context, false),
-        ),
-      ),
+      appBar: ScheduleBar(Colors.black, Theme.of(context).scaffoldBackgroundColor),
       body: SingleChildScrollView(
         child: Column(
             crossAxisAlignment: CrossAxisAlignment.start,
@@ -138,6 +128,8 @@ class _DetailState extends State<Detail> {
                     }
                 )
               ],
-            )));
+            )
+        )
+    );
   }
 }
