+++ lib/pages/Review_Ride.dart
@@ -1,6 +1,6 @@
-import 'package:carriage_rider/pages/Assistance.dart';
 import 'package:carriage_rider/pages/Ride_Confirmation.dart';
 import 'package:carriage_rider/providers/RiderProvider.dart';
+import 'package:carriage_rider/widgets/Assistance.dart';
 import 'package:flutter/material.dart';
 import 'package:carriage_rider/models/RideObject.dart';
 import 'package:provider/provider.dart';
